Add Review functionality and update User model annotation
This commit introduces functionality for handling reviews in the application. A new model, service, batch configuration, and repository have been added for Review. The User model has been updated to use the @Document and @Id annotations for compatibility with Mongo. A reviewfile path has also been added to the application.yaml and changes were made to the docker-compose.yml file.