Add Spring Boot cache support and expand description in README
A new dependency for Spring Boot's cache starter has been introduced into the pom.xml file and cache configuration was added to application.yaml, including cache names and caffeine specs. The README was expanded with additional project information and instructions. Furthermore, caching has been enabled in BatchConfiguration, and a new method to read reviews was added to ReviewService. Also implemented is a cacheable endpoint in JobLauncherController that fetches a review by ID with a delay of 5 seconds.