Update README and add Business Attributes Analysis
The README.md file has been updated with new details including a table of contents, more about the data and the project's purpose, and instructions on several aspects of the project including an enumeration of business attributes. A 'BUSSINESS-ATTRIBUTES.md' file has been added to further explain the business attributes analysis.