Update README.md
The README.md file structure has been revised to make it more intuitive and user-friendly. The 'Building and Running' section has been added to important sections at the top of the document which provides detailed instructions on how to start up the services and the application. The 'How to use the docker-compose.yml' section has been relocated under the new 'Building and Running' section.